separate output configs
