+++ pkg/config/base.go
@@ -22,6 +22,7 @@ type BaseConfig struct {
 	Logging              logger.Config      `yaml:"logging"`
 	LogLevel             string             `yaml:"log_level"`  // TODO: deprecate
 	ClusterID            string             `yaml:"cluster_id"` // Which cluster this egress belongs to
+	// BackupStorage        string             `yaml:"backup_storage"` // TODO: Files will be moved here if the upload fails
 
 	S3     *S3Config    `yaml:"s3"`
 	Azure  *AzureConfig `yaml:"azure"`
