+++ pkg/pipeline/sink/file.go
@@ -7,22 +7,21 @@ import (
 
 	"github.com/livekit/egress/pkg/config"
 	"github.com/livekit/egress/pkg/pipeline/sink/uploader"
-	"github.com/livekit/egress/pkg/types"
 	"github.com/livekit/protocol/logger"
 )
 
 type FileSink struct {
-	uploader.Uploader
+	*uploader.Uploader
 
 	conf *config.PipelineConfig
-	*config.OutputConfig
+	*config.FileConfig
 }
 
-func newFileSink(u uploader.Uploader, conf *config.PipelineConfig, out *config.OutputConfig) *FileSink {
+func newFileSink(u *uploader.Uploader, conf *config.PipelineConfig, o *config.FileConfig) *FileSink {
 	return &FileSink{
-		Uploader:     u,
-		conf:         conf,
-		OutputConfig: out,
+		Uploader:   u,
+		conf:       conf,
+		FileConfig: o,
 	}
 }
 
@@ -41,7 +40,7 @@ func (s *FileSink) Finalize() error {
 	if !s.DisableManifest {
 		manifestLocalPath := fmt.Sprintf("%s.json", s.LocalFilepath)
 		manifestStoragePath := fmt.Sprintf("%s.json", s.StorageFilepath)
-		if err = uploadManifest(s.conf, s, manifestLocalPath, manifestStoragePath); err != nil {
+		if err = uploadManifest(s.conf, s.Uploader, manifestLocalPath, manifestStoragePath); err != nil {
 			return err
 		}
 	}
@@ -61,24 +60,4 @@ func (s *FileSink) Cleanup() {
 			logger.Errorw("could not delete temp dir", err)
 		}
 	}
-}
-
-func uploadManifest(conf *config.PipelineConfig, u uploader.Uploader, localFilepath, storageFilepath string) error {
-	manifest, err := os.Create(localFilepath)
-	if err != nil {
-		return err
-	}
-
-	b, err := conf.GetManifest(types.EgressTypeFile)
-	if err != nil {
-		return err
-	}
-
-	_, err = manifest.Write(b)
-	if err != nil {
-		return err
-	}
-
-	_, _, err = u.Upload(localFilepath, storageFilepath, types.OutputTypeJSON)
-	return err
 }
