+++ pkg/pipeline/sink/uploader/backup.go
@@ -1,0 +1,1 @@
+package uploader
